Flesh out the functionality a bit more
This change improves the history trackng, adds limited support for
multiple personalities (but this not currently used).
It also makes use of the system prompt to provide some real-time
information.

A lot more things are possible. tiktoken-rs was added so the next time I
work on this I can detect when we're going to reach the max chatgpt
message size.

future work also includes using sqlite-vss to store vectors and allow
for longer conversations.